Add files via upload
This is a game using a bash script, showing how to work with a (fake) 2-dimensional array in bash. It is derived from an old DOS game called "Madness." There is score-keeping for multiple players, and you need appropriate files with r-w permissions to keep scores. The theoretical maximum score is about 12,000. A friend of mine reached almost 9,000, which is probably close to the practical limit, but there is still room at the top. The learning curve is steep.